fix Linux client destruction and header buffer sizing
Make EthernetClient safe for polymorphic deletion and verify deletion of an SSL client through its base type under ASan/UBSan. Store validated header sizes as size_t, guard undersized or unavailable buffers, and handle internal allocation failure without changing the signed constructor contract.

+++ OpenThingsFramework.h
@@ -63,7 +63,7 @@ namespace OTF {
     CLOUD_STATUS cloudStatus = NOT_ENABLED;
     uint32_t lastCloudStatusChangeTime = millis();
     char *headerBuffer = NULL;
-    int headerBufferSize = 0;
+    size_t headerBufferSize = 0;
     bool ownsHeaderBuffer = false;
 
     void webSocketEventCallback(WSEvent_t type, uint8_t *payload, size_t length);
